web_api_jwt/web_api_jwt/requirements.txt
jsonschema==4.16.0
flask
flask-restx
pyjwt
python-dotenv
oracledb
python-multipart
bcrypt
uvicorn
fastapi